Created AccountInfo. Reduced UserKey.
