+++ src/factiva/analytics/auth/__init__.py
@@ -4,7 +4,8 @@ Factiva Auth Module
 
 
 """
-__all__ = ['UserKey', 'OAuthUser']
+__all__ = ['UserKey', 'OAuthUser', 'AccountInfo']
 
 from .userkey import UserKey
 from .oauthuser import OAuthUser
+from .accountinfo import AccountInfo
